Move test with multiple deposits
